moved work to fringe search. some fixes.
