+++ algolabra/fringe/fringe_search.py
@@ -1,65 +1,88 @@
-from codecs import xmlcharrefreplace_errors
+from algolabra.fringe.doublelinkedlist import DoubleLinkedList, Node
+from algolabra.bostonmaps.map_component import read_map
 
-from fringe.doublelinkedlist import DoubleLinkedList
-
-# [0, 'Boston_0_512.map', 512, 512, 344, 85, 343, 85, 1.0]
-def children(node):
-    return [(0,0),(1,1),(2,2),(3,3)]
 
 def fringe_search(start, goal, citymap):
-    goaly, goalx = goal
-    fringe = DoubleLinkedList(*start)
+    print(f"{start=}")
+    print(f"{goal=}")
+
+    # let's make nodes
+    start_node = Node(*start)
+
+    fringe = DoubleLinkedList(node=start_node)
     cache = [[None for a in line] for line in citymap]
-    # cache = [None for i in range(len(map)*)]
-    cache[start[0]][start[1]] = (0, None, None)
-    flimit = heuristics(start, goal)
+
+    cache[start_node.y][start_node.x] = 0, None
+    flimit = heuristics(start_node, *goal)
     found = False
+    print(f"initial {flimit=}")
 
     while not found and fringe.head:
         fmin = 1000000
         for node in fringe:
             g, parent = cache[node.y][node.x]
-            f = g + heuristics(node, goal)
+            f = g + heuristics(node, *goal)
+            print(f"handling {node=}, {g=}, {parent=}, {f=}")
             if f > flimit:
                 fmin = min(f, fmin)
                 continue
-            if node.x == goal.x and node.y == goal.y:
+            if node.x == goal[0] and node.y == goal[1]:
                 print(f"found route with cost {g}")
+                print(f"{g=}, {node.x=}, {node.y=}")
+                print(f"fringe: {[(node.x, node.y) for node in fringe]}")
                 found = True
                 break
 
-            for childy, childx, cost in reversed(children(node)):
+            for x, y, cost in children(node, citymap):
+                print(f"handling child {x=},{y=}, {cost=}")
                 g_child = g + cost
-                if cache[childy][childx]:
-                    g_cached, parent = cache[childy][childx]
+                if cache[y][x]:
+                    g_cached, parent = cache[y][x]
                     if g_child >= g_cached:
                         continue
                 # TODO replace this with a doublelinkedlist function. do a search, rm if found.
                 # if child in fringe:
                 #     fringe.remove_data(child)
-                fringe.removal_version(childx, childy)
-                fringe.add_node(childx, childy, node=node)
-                cache[childy][childx] = (g_child, node.data)
-                if childx == goalx and childy == goaly:
-                    print(f"{childx=},{childy=}: {g_child=}, {node.data=}")
-                    print(f"fringe: {[(node.x,node.y) for node in fringe]}")
+                fringe.removal_version(x, y)
+                fringe.add_node(x=x, y=y, node=node)
+                cache[y][x] = g_child, (node.x, node.y)
             fringe.remove_node(node)
 
         flimit = fmin
+        print(f"updated {flimit=}")
     if found:
         route = [goal]
         while route[-1] != start:
-            indeksi = route[-1]
-            route.append(cache[indeksi])
+            x,y = route[-1]
+            route.append(cache[y][x][1])
         print(route)
         return route
 
 
-def heuristics(node, goal):
-    if node == goal:
-        return 0
-    # distance * something ideal
-    # ohkay this needs small values.
-    # that mabbe grow the closer you get?
-    # let's do a random value for starters, 1 is random enough :D
-    return 1
+def heuristics(node, goalx, goaly):
+    # undirected heuristics and cache for now
+    delta_x = node.x - goalx
+    delta_y = node.y - goaly
+
+    return max(abs(delta_x), abs(delta_y)) + (2**0.5 - 1) * min(delta_x, delta_y)
+
+def children(node, citymap):
+    """
+    Gives the valid neighbors of node in map.
+    :param node: Node
+    :param citymap: list[y][x]
+    :return: [(x,y),(x,y),...]
+    """
+    # we could just order the masks in the direction of goal node.
+    diag_cost = 2**0.5
+    masks = [(-1, -1, diag_cost), (-1, 0, 1), (-1, 1, diag_cost), (0, -1, 1), (0, 1, 1), (1, -1, diag_cost), (1, 0, 1), (1, 1, diag_cost)]
+    applied = [(mask[0] + node.x, mask[1] + node.y, mask[2]) for mask in masks]
+
+    in_area = [child for child in applied if
+               0 <= child[1] < len(citymap) and
+               0 <= child[0] < len(citymap[0])]
+    open_spaces = [loc for loc in in_area if citymap[loc[1]][loc[0]] == "." or "G"]
+    # and then reverse the dir here before returning.
+
+    # should run heuristics here and sort the children
+    return open_spaces[::-1]
